+++ ksyun/data_source_ksyun_iam_roles_test.go
@@ -1,0 +1,30 @@
+package ksyun
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+)
+
+func TestAccKsyunIAMRolesDataSource_basic(t *testing.T) {
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDataIAMRoleConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckIDExists("data.ksyun_iam_roles.roles"),
+				),
+			},
+		},
+	})
+}
+
+const testAccDataIAMRoleConfig = `
+data "ksyun_iam_roles" "roles" {
+	
+}
+`