+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.17.0(Oct 08, 2024)
+
+FEATURES:
+
+- - **New Resource:** `ksyun_iam_user` IAM用户管理
+- - **New Resource:** `ksyun_iam_role` IAM角色管理
+- - **New Resource:** `ksyun_iam_group` IAM用户组管理
+- - **New Resource:** `ksyun_iam_policy` IAM策略管理
+- - **New Data Source:** `ksyun_iam_users` IAM用户列表
+- - **New Data Source:** `ksyun_iam_roles` IAM角色列表
+- - **New Data Source:** `ksyun_iam_groups` IAM角色列表
+
 ## 1.16.3 (Aug 22, 2024)
 
 BUGFIX:
@@ -776,3 +788,26 @@ DATA SOURCES:
 
 * redis read
 * redis security group read
+
+###  IAM
+
+DATA SOURCES:
+
+* iam users read
+* iam roles read
+* iam groups read
+
+RESOURCES:
+
+* iam user create
+* iam user read
+* iam user delete
+* iam role create
+* iam role read
+* iam role delete
+* iam group create
+* iam group read
+* iam group delete
+* iam policy create
+* iam policy read
+* iam policy delete